step2 수정
main.swift 와 Step2.siwft 파일 분리
findDuplicateNumbersInTwoLottoNumbers 함수와 printDuplicateNumbers 함수로 분리
